+++ R/dplyr-join.R
@@ -1,3 +1,20 @@
+rename_join_tsibble <- function(x, y, by = NULL, suffix = c(".x", ".y")) {
+  intersect_names <- intersect(names(x), names(y))
+  if (is_null(by)) {
+    by <- intersect_names
+  }
+  common_names <- setdiff(intersect_names, by)
+  if (has_length(common_names)) {
+    names <- paste0(common_names, suffix[1])
+    names(common_names) <- names
+    val_names <- tidyselect::vars_rename(names(x), !!! common_names)
+    x <- rename_index(x, val_names) %>% 
+      rename_index2(val_names) %>% 
+      rename_key(val_names)
+  }
+  x
+}
+
 #' @importFrom dplyr left_join
 #' @export
 dplyr::left_join
@@ -11,7 +28,8 @@ left_join.tbl_ts <- function(
   x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...
 ) {
   tbl <- NextMethod()
-  update_meta(tbl, x, ordered = is_ordered(x), validate = FALSE)
+  x <- rename_join_tsibble(x, y, by = by, suffix = suffix)
+  update_meta(tbl, x, ordered = is_ordered(x), validate = TRUE)
 }
 
 #' @importFrom dplyr right_join
@@ -24,6 +42,7 @@ right_join.tbl_ts <- function(
   x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...
 ) {
   tbl <- NextMethod()
+  x <- rename_join_tsibble(x, y, by = by, suffix = suffix)
   update_meta(tbl, x, ordered = is_ordered(x), validate = TRUE)
 }
 
@@ -37,6 +56,7 @@ inner_join.tbl_ts <- function(
   x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...
 ) {
   tbl <- NextMethod()
+  x <- rename_join_tsibble(x, y, by = by, suffix = suffix)
   update_meta(tbl, x, ordered = is_ordered(x), validate = TRUE)
 }
 
@@ -50,6 +70,7 @@ full_join.tbl_ts <- function(
   x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...
 ) {
   tbl <- NextMethod()
+  x <- rename_join_tsibble(x, y, by = by, suffix = suffix)
   update_meta(tbl, x, ordered = is_ordered(x), validate = TRUE)
 }