+++ R/update.R
@@ -13,6 +13,10 @@ by_row <- function(FUN, .data, ordered = TRUE, interval = NULL, ...,
 update_meta <- function(
   new, old, ordered = TRUE, interval = NULL, validate = FALSE
 ) {
+  if (validate) {
+    retain_tsibble(new, key = key(old), index = index(old))
+  }
+  validate <- TRUE || validate
   restore_index_class(build_tsibble(
     new, key = key(old), index = !! index(old), index2 = !! index2(old),
     regular = is_regular(old), ordered = ordered, interval = interval,