+++ NEWS.md
@@ -11,6 +11,7 @@
 * Fixed bugs in `as_tsibble.ts()` for monthly series that not start at January. (#89)
 * `guess_frequency.yearweek()` returns 52.18 for more accurate weekly representation, instead of 52.
 * `n()` now can be called in `slice.tbl_ts()`. (#95)
+* Fixed `*_join()` for not finding key or index when `by` is specified. (#102)
 
 ## Misc